\name{ex.a}
\alias{ex.a}
\encoding{UTF-8}
\docType{data}
\title{
Site 1
}
\description{
Example of a cubic trend (Site 1, Jamali et al. 2014)
}
\usage{data("ex.a")}
\format{
  The format is:
 num [1:25] 0.477 0.526 0.52 0.571 0.554 ...
}
\source{
Jamali S, Seaquist J, Eklundh L, Ardö J (2014). Automated mapping of vegetation trends with polynomials using NDVI imagery over the Sahel. Remote Sensing of Environment, 141, 79-89. http://dx.doi.org/10.1016/j.rse.2013.10.019
}
\examples{
data(ex.a)
## maybe str(ex.a) ; plot(ex.a) ...
}
\keyword{datasets}
