\name{ex.b}
\alias{ex.b}
\encoding{UTF-8}
\docType{data}
\title{
Site 2
}
\description{
Example of a cubic trend (Site 2, Jamali et al. 2014)
}
\usage{data("ex.b")}
\format{
  The format is:
 num [1:25] 0.254 0.267 0.184 0.22 0.208 ...
}
\source{
Jamali S, Seaquist J, Eklundh L, Ardö J (2014). Automated mapping of vegetation trends with polynomials using NDVI imagery over the Sahel. Remote Sensing of Environment, 141, 79-89. http://dx.doi.org/10.1016/j.rse.2013.10.019
}
\examples{
data(ex.b)
## maybe str(ex.b) ; plot(ex.b) ...
}
\keyword{datasets}
