\name{ex.c}
\alias{ex.c}
\encoding{UTF-8}
\docType{data}
\title{
Site 3
}
\description{
Example of a concealed trend (Site 3, Jamali et al. 2014)
}
\usage{data("ex.c")}
\format{
  The format is:
 num [1:25] 0.712 0.726 0.736 0.723 0.739 ...
}
\source{
Jamali, S, Seaquist, J, Eklundh, L, & Ardö, J 2014, 'Automated mapping of vegetation trends with polynomials using NDVI imagery over the Sahel', Remote Sensing Of Environment, 141, pp. 79-89, ScienceDirect, EBSCOhost.
}
\examples{
data(ex.c)
## maybe str(ex.c) ; plot(ex.c) ...
}
\keyword{datasets}
