\name{ex.d}
\alias{ex.d}
\encoding{UTF-8}
\docType{data}
\title{
Site 4
}
\description{
Example of a concealed trend with cubic form (Site 4, Jamali et al. 2014)
}
\usage{data("ex.d")}
\format{
  The format is:
 num [1:25] 0.6 0.54 0.447 0.478 0.457 ...
}
\source{
Jamali S, Seaquist J, Eklundh L, Ardö J (2014). Automated mapping of vegetation trends with polynomials using NDVI imagery over the Sahel. Remote Sensing of Environment, 141, 79-89. http://dx.doi.org/10.1016/j.rse.2013.10.019
}
\examples{
data(ex.d)
## maybe str(ex.d) ; plot(ex.d) ...
}
\keyword{datasets}
