\name{ex.e}
\alias{ex.e}
\encoding{UTF-8}
\docType{data}
\title{
Site 5
}
\description{
Example of a quadratic trend (Site 5, Jamali et al. 2014)
}
\usage{data("ex.e")}
\format{
  The format is:
 num [1:25] 0.26 0.278 0.196 0.224 0.261 ...
}
\source{
Jamali S, Seaquist J, Eklundh L, Ardö J (2014). Automated mapping of vegetation trends with polynomials using NDVI imagery over the Sahel. Remote Sensing of Environment, 141, 79-89. http://dx.doi.org/10.1016/j.rse.2013.10.019
}
\examples{
data(ex.e)
## maybe str(ex.e) ; plot(ex.e) ...
}
\keyword{datasets}
