\name{ex.g}
\alias{ex.g}
\encoding{UTF-8}
\docType{data}
\title{
Site 7
}
\description{
Example of a concealed trend with quadratic form (Site 7, Jamali et al. 2014)
}
\usage{data("ex.g")}
\format{
  The format is:
 num [1:25] 0.147 0.153 0.104 0.123 0.15 ...
}
\source{
Jamali S, Seaquist J, Eklundh L, Ardö J (2014). Automated mapping of vegetation trends with polynomials using NDVI imagery over the Sahel. Remote Sensing of Environment, 141, 79-89. http://dx.doi.org/10.1016/j.rse.2013.10.019
}
\examples{
data(ex.g)
## maybe str(ex.g) ; plot(ex.g) ...
}
\keyword{datasets}
