\name{ex.h}
\alias{ex.h}
\encoding{UTF-8}
\docType{data}
\title{
Site 8
}
\description{
Example of a concealed trend (Site 8, Jamali et al. 2014)
}
\usage{data("ex.h")}
\format{
  The format is:
 num [1:25] 0.813 0.809 0.753 0.792 0.727 ...
}
\source{
Jamali, S, Seaquist, J, Eklundh, L, & Ardö, J 2014, 'Automated mapping of vegetation trends with polynomials using NDVI imagery over the Sahel', Remote Sensing Of Environment, 141, pp. 79-89, ScienceDirect, EBSCOhost.
}
\examples{
data(ex.h)
## maybe str(ex.h) ; plot(ex.h) ...
}
\keyword{datasets}
