\name{ex.k}
\alias{ex.k}
\encoding{UTF-8}
\docType{data}
\title{
Site 9
}
\description{
Example of a linear trend (Site 9, Jamali et al. 2014)
}
\usage{data("ex.k")}
\format{
  The format is:
 num [1:25] 0.382 0.373 0.228 0.385 0.271 ...
}
\source{
Jamali S, Seaquist J, Eklundh L, Ardö J (2014). Automated mapping of vegetation trends with polynomials using NDVI imagery over the Sahel. Remote Sensing of Environment, 141, 79-89. http://dx.doi.org/10.1016/j.rse.2013.10.019
}
\examples{
data(ex.k)
## maybe str(ex.k) ; plot(ex.k) ...
}
\keyword{datasets}
