\name{ex.m}
\alias{ex.m}
\encoding{UTF-8}
\docType{data}
\title{
Site 10
}
\description{
Example of a linear trend (Site 10, Jamali et al. 2014)
}
\usage{data("ex.m")}
\format{
  The format is:
 num [1:25] 0.781 0.738 0.795 0.871 0.736 ...
}
\source{
Jamali S, Seaquist J, Eklundh L, Ardö J (2014). Automated mapping of vegetation trends with polynomials using NDVI imagery over the Sahel. Remote Sensing of Environment, 141, 79-89. http://dx.doi.org/10.1016/j.rse.2013.10.019
}
\examples{
data(ex.m)
## maybe str(ex.m) ; plot(ex.m) ...
}
\keyword{datasets}
