\name{ex.n}
\alias{ex.n}
\encoding{UTF-8}
\docType{data}
\title{
Site 11
}
\description{
Example of a no-trend (Site 11, Jamali et al. 2014)
}
\usage{data("ex.n")}
\format{
  The format is:
 num [1:25] 0.567 0.625 0.61 0.589 0.569 ...
}
\source{
Jamali, S, Seaquist, J, Eklundh, L, & Ardö, J 2014, 'Automated mapping of vegetation trends with polynomials using NDVI imagery over the Sahel', Remote Sensing Of Environment, 141, pp. 79-89, ScienceDirect, EBSCOhost.
}
\examples{
data(ex.n)
## maybe str(ex.n) ; plot(ex.n) ...
}
\keyword{datasets}
