\name{ex.d}
\alias{ex.d}
\encoding{UTF-8}
\docType{data}
\title{
Site 5
}
\description{
Example of a quadratic trend (Site 5, Jamali et al. 2014)
}
\usage{data("ex.d")}
\format{
  The format is:
 num [1:25] 0.26 0.278 0.196 0.224 0.261 ...
}
\source{
Jamali, S, Seaquist, J, Eklundh, L, & Ardö, J 2014, 'Automated mapping of vegetation trends with polynomials using NDVI imagery over the Sahel', Remote Sensing Of Environment, 141, pp. 79-89, ScienceDirect, EBSCOhost.
}
\examples{
data(ex.d)
## maybe str(ex.d) ; plot(ex.d) ...
}
\keyword{datasets}
