\name{ex.e}
\alias{ex.e}
\encoding{UTF-8}
\docType{data}
\title{
Site 4
}
\description{
Example of a concealed trend (Site 4, Jamali et al. 2014)
}
\usage{data("ex.e")}
\format{
  The format is:
 num [1:25] 0.6 0.54 0.447 0.478 0.457 ...
}
\source{
Jamali, S, Seaquist, J, Eklundh, L, & Ardö, J 2014, 'Automated mapping of vegetation trends with polynomials using NDVI imagery over the Sahel', Remote Sensing Of Environment, 141, pp. 79-89, ScienceDirect, EBSCOhost.
}
\examples{
data(ex.e)
## maybe str(ex.e) ; plot(ex.e) ...
}
\keyword{datasets}
