\name{ex.f}
\alias{ex.f}
\encoding{UTF-8}
\docType{data}
\title{
Site 6
}
\description{
Example of a quadratic trend (Site 6, Jamali et al. 2014)
}
\usage{data("ex.f")}
\format{
  The format is:
 num [1:25] 0.327 0.259 0.212 0.298 0.248 ...
}
\source{
Jamali, S, Seaquist, J, Eklundh, L, & Ardö, J 2014, 'Automated mapping of vegetation trends with polynomials using NDVI imagery over the Sahel', Remote Sensing Of Environment, 141, pp. 79-89, ScienceDirect, EBSCOhost.
}
\examples{
data(ex.f)
## maybe str(ex.f) ; plot(ex.f) ...
}
\keyword{datasets}
